requirements.txt
numpy
pandas
Pillow
scikit-image
scipy
tqdm
